add installer

+++ configure.sh
@@ -1,34 +1,80 @@
 #!/usr/bin/env bash
-#
+
+
+ppa_exists() {
+    ppa=$(find /etc/apt/ -name *.list | xargs cat | grep  ^[[:space:]]*deb | awk -F'[/:]+' '{ sub("^www", "", $2); print $3 }' | grep $1)
+    if [ ! -z "${ppa}" ]; then
+        echo "1"
+        return
+    fi
+
+    echo "0"
+}
+
 install_repositories() {
     echo "install repositories"
 
-    echo "- install emacs repository"
-    sudo add-apt-repository ppa:kelleyk/emacs
-    sudo apt update
+    exists=$(ppa_exists "kelleyk")
+    if [ "$exists" == "0" ]; then
+        echo "- install emacs repository"
+        sudo add-apt-repository ppa:kelleyk/emacs -y > /dev/null 2>&1
+        sudo apt-get update -qq
+    else
+        echo "- emacs-ppa repository: already added"
+    fi
 
-    echo "- install neovim repository"
-    sudo add-apt-repository ppa:neovim-ppa/unstable
-    sudo apt update
+    exists=$(ppa_exists "neovim-ppa")
+    if [ "$exists" == "0" ]; then
+        echo "- install neovim repository"
+        sudo add-apt-repository ppa:neovim-ppa/unstable -y > /dev/null 2>&1
+        sudo apt-get update -qq
+    else
+        echo "- neovim-ppa: already added"
+    fi
 
-    echo "- install peek repository"
-    sudo add-apt-repository ppa:peek-developers/stable
-    sudo apt update
+    exists=$(ppa_exists "peek-developers")
+    if [ "$exists" == "0" ]; then
+        echo "- install peek repository"
+        sudo add-apt-repository ppa:peek-developers/stable -y > /dev/null 2>&1
+        sudo apt-get update -qq
+    else
+        echo "- peek-ppa: already added"
+    fi
 
     [[ -f "/etc/os-release" ]] && ID="$(awk -F= ' /^ID=/ { gsub("\"", ""); print $2 } ' /etc/os-release)"
-    if [ "$ID" = "ubuntu" ];
-        echo "- install alacritty repository"
-        sudo add-apt-repository ppa:mmstick76/alacritty
-        sudo apt update
+    if [ "$ID" == "ubuntu" ]; then
+        exists=$(ppa_exists "aslatter")
+        if [ "$exists" == "0" ]; then
+            echo "- install alacritty repository"
+            sudo add-apt-repository ppa:aslatter/ppa -y > /dev/null 2>&1
+            sudo apt-get update -qq
+        else
+            echo "- alacritty-ppa: already added"
+        fi
+    fi
+}
+
+install_docker_repository() {
+    echo "install docker repository"
+
+    if [ ! -f "/etc/apt/keyrings/docker.gpg" ]; then
+        sudo install -m 0755 -d /etc/apt/keyrings
+        curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+        sudo chmod a+r /etc/apt/keyrings/docker.gpg
+
+        echo "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu "$(. /etc/os-release && echo "$VERSION_CODENAME")" stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+
+        sudo apt-get update -qq
+    else
+        echo "- docker: already installed"
     fi
 }
 
 install_packages() {
     echo "- install packages"
     dir=$(dirname $0)
-    filename=${dir}/packages.txt
-
-    echo "sudo apt install " $(grep -vE "^\s*#" ${filename}  | tr "\n" " ")
+    filename=${dir}/$1
+    sudo apt-get install -qq $(grep -vE "^\s*#" ${filename}  | tr "\n" " ") -y
 }
 
 install_fonts() {
@@ -47,9 +93,15 @@ install_fonts() {
         fontpkg=${data##*/}
         file="${fontpkg%.*}"
         echo "installing ${file}"
-        wget -q $data -P /tmp
-        unzip /tmp/$fontpkg -d "$fonts_dir/$file" > /dev/null
-        rm -rf /tmp/$fontpkg
+
+        if [ ! -d "$fonts_dir/$file" ]; then
+            wget -q $data -P /tmp
+            unzip /tmp/$fontpkg -d "$fonts_dir/$file" > /dev/null
+            echo "${file}: installed"
+            rm -rf /tmp/$fontpkg
+        else
+            echo "${file}: alredy installed"
+        fi
     done < $filename
 
     echo "updating font cache"
@@ -58,18 +110,20 @@ install_fonts() {
 }
 
 install_startship() {
-    if [ ! -x starship ]; then
+    if [ -x starship ]; then
         echo "starship: already done"
         return
     fi
 
-    curl -sS https://starship.rs/install.sh | sh
+    wget -q https://starship.rs/install.sh -P /tmp
+    chmod 0755 "/tmp/install.sh"
+    /tmp/install.sh --yes > /dev/null
     echo "starship: installed"
 }
 
 install_tmux_plugins() {
 
-    if [ -d $HOME/.tmux/plugins/tpm ]; then
+    if [ -d "$HOME/.tmux/plugins/tpm" ]; then
         echo "tmux plugins: already done"	
         return
     fi
@@ -81,7 +135,7 @@ install_tmux_plugins() {
 install_asdf() {
     echo "install asdf"
 
-    if [ -d $HOME/.asdf ]; then
+    if [ -d "$HOME/.asdf" ]; then
         echo "asdf: already done"
         return
     fi
@@ -90,8 +144,66 @@ install_asdf() {
     echo "asdf: installed"
 }
 
-install_packages
-install_fonts
-install_startship
-install_tmux_plugins
-install_asdf
+install_rust() {
+    echo "install rust"
+
+    if [ -d "$HOME/./.rustup/toolchains" ]; then
+        echo "rust: already done"
+        return
+    fi
+
+    wget -q https://sh.rustup.rs -O /tmp/sh.rustup.rs
+    chmod 0755 "/tmp/sh.rustup.rs"
+    /tmp/sh.rustup.rs -y > /dev/null
+    echo "rust: installed"
+}
+
+install_compose() {
+    sudo groupadd docker
+
+    sudo usermod -aG docker $USER
+    sudo curl -sSL https://github.com/docker/compose/releases/download/v2.20.2/docker-compose-linux-x86_64 -o /usr/local/bin/docker-compose
+    sudo chmod 755 /usr/local/bin/docker-compose
+    sudo chown root:docker /usr/local/bin/docker-compose
+}
+
+usage() {
+    echo -e "\nUsage:\n"
+    echo "f - to install and full configuration"
+}
+
+check_user() {
+    if [ -z $USER ]; then
+        echo "Running in docker variable USER is not set"
+        echo 'USER=$(whoami) ./configure.sh option '
+        exit 1;
+    fi
+}
+
+main() {
+    check_user
+    case $1 in
+        f)
+            install_repositories
+            install_docker_repository
+            install_packages "full_packages.txt"
+            install_fonts
+            install_startship
+            install_tmux_plugins
+            install_asdf
+            install_rust
+            install_compose
+            ;;
+        m) 
+            install_repositories
+            install_packages "min_packages.txt"
+            install_startship
+            install_tmux_plugins
+            install_asdf
+            install_rust
+            ;;
+         *) usage; exit 666;
+    esac
+}
+
+main "$1"
